cleaned up handling of storage specification; static file handling in testing
