+++ pyensemble/urls.py
@@ -14,7 +14,6 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.conf import settings
-from django.conf.urls.static import static
 
 from django.contrib import admin
 from django.contrib.auth import views as auth_views
@@ -112,7 +111,3 @@ urlpatterns = [
     path('integrations/', include(integrations_urls, namespace='integrations')),
     path('integrity/', include(integrity_patterns)),
 ]
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL,
-                          document_root=settings.MEDIA_ROOT)
